+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.0
 
 - Integration with EasySAM
+- Added export decorator
 
 # 1.0.0